+++ tx-backend/src/main/java/org/eclipse/tractusx/traceability/common/properties/TraceabilityProperties.java
@@ -51,4 +51,5 @@ public class TraceabilityProperties {
     private String regularApiKey;
     private String irsBase;
     private String discoveryType;
+    private String technicalServiceApiKey;
 }